Reorganize profiler functionality
